Create a GStreamerRender helper object
This object is in charge of two tasks:

- Create the video sink
- Create the frames to push to the Renders

The purpose of this design is to handle GL by another create (in order
to workaround
https://github.com/servo/servo/pull/22944#issuecomment-468827665), and
that other object will proxy these tasks if available in the target
OS.

This patch removes temporarly the GL support in player.
