+++ backends/gstreamer/Cargo.toml
@@ -52,9 +52,6 @@ version = "0.13"
 [dependencies.gstreamer-video]
 version = "0.13"
 
-[dependencies.gstreamer-gl]
-version = "0.13"
-
 [dependencies.ipc-channel]
 version = "0.11"
