+++ backends/gstreamer/lib.rs
@@ -13,7 +13,6 @@ extern crate glib;
 extern crate gstreamer as gst;
 extern crate gstreamer_app as gst_app;
 extern crate gstreamer_audio as gst_audio;
-extern crate gstreamer_gl as gst_gl;
 extern crate gstreamer_player as gst_player;
 extern crate gstreamer_sdp as gst_sdp;
 extern crate gstreamer_video as gst_video;
@@ -35,6 +34,7 @@ pub mod media_capture;
 pub mod media_stream;
 mod media_stream_source;
 pub mod player;
+mod render;
 mod source;
 pub mod webrtc;
